weather_chatbot/backend/pyproject.toml
[project]
name = "techstack-practise"
version = "0.1.0"
description = "Agentic AI"
requires-python = ">=3.12"

authors = [
    { name = "Sahishnu G (Intern) :: Kshema" }
]

# Core agentic + web stack
dependencies = [
    # Web API
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",

    # HTTP + env
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",

    # LangGraph / LangChain / Ollama
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-core>=0.2.0",
    "langchain-ollama>=0.1.0",

    # Data models (FastAPI uses Pydantic v2)
    "pydantic>=2.5.0",

    # Optional: if you actually need PyTorch; otherwise remove
    "langfuse<3.0.0"
]

